+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # History of Changes
 
+## v2024.11.3
+
+- Core.Reader can now determine voltage_step from source-file
+- HarvesterPRUConfig.from_vhrv() needs voltage_step IF input is IVCurve for emulation
+  - same for init() of VirtualSourceModel
+  - this fixes a bug that could ruin emulations with ivcurves (#73)
+
 ## v2024.11.2
 
 - adapt fixtures to recent testbed-restructure